+++ ui-src/components/ExportForm.tsx
@@ -1,45 +1,20 @@
-import {
-  Button,
-  Muted,
-  SegmentedControl,
-  type SegmentedControlOption
-} from '@create-figma-plugin/ui';
-import type { JSX, TargetedEvent } from 'preact';
-import { useEffect } from 'preact/hooks';
+import { Button } from '@create-figma-plugin/ui';
+import type { JSX } from 'preact';
 import { FormProvider, useForm } from 'react-hook-form';
 
-import { ExternalLibrariesFieldSet } from '@ui/components/ExternalLibrariesFieldSet';
+import { DesignExportControls } from '@ui/components/DesignExportControls';
 import { Stack } from '@ui/components/Stack';
 import { TwoColumnLayout } from '@ui/components/TwoColumnLayout';
 import { type FormValues, useFigmaContext } from '@ui/context';
-import type { ExportScope } from '@ui/types';
-
-const scopeOptions: SegmentedControlOption[] = [
-  { value: 'all', children: 'All pages' },
-  { value: 'current', children: 'Current page' }
-];
 
 export const ExportForm = (): JSX.Element => {
-  const { cancel, exportPenpot, exportScope, exportLibraries, setExportScope, editorType } =
-    useFigmaContext();
+  const { cancel, exportPenpot, editorType } = useFigmaContext();
   const isSlidesEditor = editorType === 'slides';
   const methods = useForm<FormValues>({
     defaultValues: {
       externalLibraries: []
     }
   });
-
-  useEffect(() => {
-    if (exportLibraries.length > 0) {
-      methods.reset({
-        externalLibraries: exportLibraries.map(name => ({ name, uuid: '' }))
-      });
-    }
-  }, [exportLibraries]);
-
-  const handleScopeChange = (event: TargetedEvent<HTMLInputElement>): void => {
-    setExportScope(event.currentTarget.value as ExportScope);
-  };
 
   const leftColumn = (
     <Stack space="medium">
@@ -74,28 +49,7 @@ export const ExportForm = (): JSX.Element => {
     </Stack>
   );
 
-  const rightColumn = (
-    <Stack space="medium">
-      {!isSlidesEditor && (
-        <Stack space="xsmall">
-          <strong style={{ fontSize: 13 }}>Export scope</strong>
-          <div style={{ width: 'fit-content' }}>
-            <SegmentedControl
-              options={scopeOptions}
-              value={exportScope}
-              onChange={handleScopeChange}
-            />
-          </div>
-          <Muted>
-            For large documents, try exporting the current page first to test, then do a full
-            export.
-          </Muted>
-        </Stack>
-      )}
-
-      {!isSlidesEditor && <ExternalLibrariesFieldSet />}
-    </Stack>
-  );
+  const rightColumn = isSlidesEditor ? null : <DesignExportControls />;
 
   return (
     <FormProvider {...methods}>
